Telemetry sources are published Updated patches and build system

+++ src/qoi_import.cpp
@@ -5,7 +5,7 @@ PackedStringArray QOIImport::_get_recognized_extensions() const {
 	return PackedStringArray(TypedArray<String>::make("qoi"));
 }
 
-Error QOIImport::_load_image(const Ref<Image> &image, const Ref<FileAccess> &fileaccess, BitField<ImageFormatLoader::LoaderFlags> flags, double scale) {
+Error QOIImport::_load_image(const Ref<Image> &image, const Ref<FileAccess> &fileaccess, BitField<ImageFormatLoader::LoaderFlags> flags, float scale) {
 	PackedByteArray data = fileaccess->get_buffer(fileaccess->get_length());
 	ERR_FAIL_COND_V_MSG(fileaccess->get_error() != Error::OK, fileaccess->get_error(), "Failed to read data from file: " + fileaccess->get_path());
