+++ src/qoi_import.h
@@ -15,5 +15,5 @@ protected:
 
 public:
 	virtual PackedStringArray _get_recognized_extensions() const override;
-	virtual Error _load_image(const Ref<Image> &image, const Ref<FileAccess> &fileaccess, BitField<ImageFormatLoader::LoaderFlags> flags, double scale) override;
+	virtual Error _load_image(const Ref<Image> &image, const Ref<FileAccess> &fileaccess, BitField<ImageFormatLoader::LoaderFlags> flags, float scale) override;
 };
